fix(#1131): relocate ipam_step_up_validate_return_to to lib
Pass-A smoke against the deployed v3.27.3 Test SQLite instance found a
fatal during the OIDC step-up replay flow:

  Fatal error: Cannot redeclare function to_str() (previously declared
  in /var/www/html/testing/ipam/init.php:154) in
  /var/www/html/testing/ipam/init.php on line 154

Root cause: the first cut of sudo_replay.php imported the validator
via `require_once __DIR__ . '/step_up.php'`. step_up.php is a PAGE
entry point — it does `require __DIR__ . '/init.php'` at the top — so
loading it from another page-entry-point caused init.php to execute
twice in the same request and redeclare every helper at file-scope
(to_str, to_int, current_user, …).

Same bug also fired on settings_reveal.php's "view OIDC client secret"
flow because it follows the same redirect chain (step_up.php → OIDC
reauth → sudo_replay.php).

Fix: move ipam_step_up_validate_return_to() from step_up.php (page)
into lib/auth_step_up.php (library, no init.php require). step_up.php
now imports the function from the lib via require_once. sudo_replay.php
drops the offending step_up.php require entirely.

Regression test at tests/SudoReplayInitOnceTest.php enforces both
invariants at the source level so the bad pattern can't sneak back in.

Local gate: phpunit (969 OK), phpstan L9.
